feature: Add ffmpeg HLS streaming support
- Adds a function to create an HLS stream using ffmpeg. It returns an
rxjs Observable that emits progress updates and errors. It is a cold
observable, meaning that the ffmpeg process is not started until the
observable is subscribed to. This makes it useful to describe a
re-usable HLS streaming process that can be started and stopped by
subscribing and unsubscribing to the observable.