chore: update jest coverage config for new module
Currently, @nx/node:library does not generate a correct jest.config.ts
when `swc` is selected as the compiler. It seems the default jest
coverage provider does not understand the `swc` output very well, so
we need to set the `v8` coverage provider instead to avoid issues
with lines reported as uncovered that are actually covered.