feature: poc for live endless audio stream plugin
Implemented core logic for a live streaming endless audio playlist as a
fastify plugin that statically serves HLS stream files. Includes a
Playlist class with efficient seek logic and property-based tests,
integration of ffmpeg-based HLS streaming and playlist file generation,
and utilities to manage playlsit metadata files.

Implemented a separate start-instant plugin for fastify to handle
computing and storing an absolute "start time" for the stream

Note: This is an initial POC; further testing, documentation, and
efactoring are needed