fix: added test audio files for server development
Added test audio files to the server project to be used for development
and testing purposes.

Added a new environment file to point to the new audio file directory
for development and test purposes.

Note: this folder and configuration can be overridden by using a local
env file, something like `.serve.development.local.env`. This is useful
for more extensive manual testing, but care shopuld be taken not to
depend on the local environment file for automated tests or CI/CD
(including the any tests in the server-e2e app)