feature: dockerize server
Add Docker support for the server application via a dockerfile and
nx targets to build and run the server in docker that hook into the nx
lifecycle. You can run the server in docker with
`nx docker-run server --inputDir=/some/dir/with/wavs`. The input dir
does need to be absolute for Docker to bind it correctly.