test: update server-e2e to use dockerized server
Added new nx targets for `docker-run` and `e2e` to the server-e2e
project that run the server in a docker container before running the e2e
tests. Also moved the test audio files to the server-e2e project to
bring them closer to the tests that use them.

Updated the global teardown script to stop the docker container after
the tests are done, but it's messy, and we might want to consider
defining it with docker compose in the future.