chore: move project assets, add deps and license
Added (OpenArt 1.0) LICENSE.md and updated package.json reference to it.
This license seems to be somewhat defunct, so it is subject to change.

Moved all project assets to the `assets` directory. This includes the
favicon files, test audio files, and generated documentation. It is
also the default target of the new `sync-audio` script, which downloads
audio files from google cloud storage to be streamed by the server app.

Added `sync-audio` script to package.json for syncing audio assets from
Google Cloud Storage

Added @google-cloud/pino-logging-gcp-config dependency to ensure prod
logs are compliant with Google Cloud Logging structure

Added eslint-plugin-jest to enhance linting in jest tests.
Unfortunately, the `@fast-check/jest` wrapper library interferes with
some linting rules that are looking for enclosing `it` blocks, which
that library replaces for convenience. Some configuration was added to
eslint.config.mjs to mitigate this issue for some rules.

Updated README.md to reflect new local dev env requirements, the gcloud
and docker tools.