chore: update server docker setup for google cloud
Set Dockerfile port to 80 so that http will map automatically on GCP VM.

Update docker-build and docker-run targets for the server app to
tag the image appropriately (:latest and :test) for production and test
builds and to use the correct audio directory in `assets`

Update server-e2e test setup and teardown to use the latest ports and
docker kill commands. Be warned, the e2e test teardown script will
kill all running containers, so that may need to change if future tests
run additional containers.