cicd: add gha workflow for deployment to gcp vm
Added a GitHub Actions workflow for deploying the server to a Google
Cloud Platform VM. This workflow includes steps for checking out the
code, building the server docker image, publishing the image to github
container registry, and deploying the server to the VM using the google
cloud sdk. It also sets the VM startup script to download audio files
from a Google Cloud Storage bucket and bind those files to the server
container on startup.